bpf_helpers: make *all* eBPF programs return long not int
BPF is a 64-bit architecture, so this makes more sense:
int is 32-bits, while long is 64-bits.

In practice this doesn't do much, but it does make sure
the top 32-bits of the 64-bit R0 register are well defined.

I think this might matter when we try to introduce function calls,
as per Linux Kernel eBPF uapi we have (for example):

  long (*callback_fn)(struct bpf_map *map, const void *key, void *value, void *ctx);
  long bpf_for_each_map_elem(struct bpf_map *map, void *callback_fn, void *callback_ctx, u64 flags);

Test: TreeHugger
Flag: EXEMPT mainline

+++ bpf/progs/include/bpf_helpers.h
@@ -549,7 +549,7 @@ static int (*bpf_trace_printk)(const char* fmt, int fmt_size, ...) = (void*) BPF
         .pin_prefix = "/sys/fs/bpf/" pindir "/prog_" BPF_OBJ_NAME "_",                   \
     };                                                                                   \
     SECTION(#TYPE "/" NAME)                                                              \
-    int the_prog
+    long the_prog
 
 #define DEFINE_BPF_PROG_KVER_RANGE_OPT(TYPE, NAME, prog_uid, prog_gid, the_prog, min_kv, max_kv, \
                                        opt)                                                      \
